logou check

+++ app/src/main/java/example/com/tourismapp2/ViewAllPlaces.java
@@ -20,6 +20,7 @@ import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.database.DataSnapshot;
@@ -29,7 +30,9 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 import com.squareup.picasso.Picasso;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 
 import example.com.tourismapp2.classpack.places_details;
 
@@ -59,12 +62,46 @@ EditText searchEt;
             @Override
             public void onClick(View view) {
                 SharedPreferences sharedPreference=getActivity().getSharedPreferences("mypref", Context.MODE_PRIVATE);
+
                 SharedPreferences.Editor editor = sharedPreference.edit();
-                editor.clear();
-                editor.apply();
-                getActivity().finish();
-                Intent intent=new Intent(getActivity(),Login_Signup.class);
-                startActivity(intent);
+                long beforeTime=sharedPreference.getLong("time",0);
+                long currenttime=new Date().getTime();
+                long timeDiff=currenttime-beforeTime;
+                SimpleDateFormat format=new SimpleDateFormat("hh-mm");
+                String value= format.format(timeDiff);
+                System.out.println(value);
+
+                boolean task1,task2;
+                task1=sharedPreference.getBoolean("TASK1",false);
+                task2=sharedPreference.getBoolean("TASK2",false);
+                boolean flag=false;
+                if(task1)
+                {
+                    if(!task2)
+                    {
+                        Toast.makeText(getActivity(), "No Place added in PLanner", Toast.LENGTH_SHORT).show();
+
+
+                    }
+                    else
+                    {
+                        flag=true;
+                    }
+
+                }
+                else {
+                    Toast.makeText(getActivity(), "No Place addded to favourites", Toast.LENGTH_SHORT).show();
+
+                }
+
+
+                if(flag) {
+                    editor.clear();
+                    editor.apply();
+                    getActivity().finish();
+                    Intent intent = new Intent(getActivity(), Login_Signup.class);
+                    startActivity(intent);
+                }
