+++ app/src/main/java/example/com/tourismapp2/View_Place_detail.java
@@ -31,7 +31,9 @@ import com.squareup.picasso.Picasso;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 
 import example.com.tourismapp2.classpack.HotelDetail;
 import example.com.tourismapp2.classpack.places_details;
@@ -172,13 +174,46 @@ public class View_Place_detail extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 SharedPreferences sharedPreference=getSharedPreferences("mypref", Context.MODE_PRIVATE);
-                SharedPreferences.Editor editor = sharedPreference.edit();
-                editor.clear();
-                editor.apply();
-             finish();
-                Intent intent=new Intent(View_Place_detail.this,Login_Signup.class);
-                startActivity(intent);
 
+                SharedPreferences.Editor editor = sharedPreference.edit();
+                long beforeTime=sharedPreference.getLong("time",0);
+                long currenttime=new Date().getTime();
+                long timeDiff=currenttime-beforeTime;
+                SimpleDateFormat format=new SimpleDateFormat("hh-mm");
+                String value= format.format(timeDiff);
+                System.out.println(value);
+
+                boolean task1,task2;
+                task1=sharedPreference.getBoolean("TASK1",false);
+                task2=sharedPreference.getBoolean("TASK2",false);
+                boolean flag=false;
+                if(task1)
+                {
+                    if(!task2)
+                    {
+                        Toast.makeText(View_Place_detail.this, "No Place added in PLanner", Toast.LENGTH_SHORT).show();
+
+
+                    }
+                    else
+                    {
+                        flag=true;
+                    }
+
+                }
+                else {
+                    Toast.makeText(View_Place_detail.this, "No Place addded to favourites", Toast.LENGTH_SHORT).show();
+
+                }
+
+
+                if(flag) {
+                    editor.clear();
+                    editor.apply();
+                    finish();
+                    Intent intent = new Intent(View_Place_detail.this, Login_Signup.class);
+                    startActivity(intent);
+                }
 
 
             }
@@ -287,6 +322,10 @@ public class View_Place_detail extends AppCompatActivity {
                                 finish();
                                 startActivity(backIntent);
                                 Toast.makeText(View_Place_detail.this, "Place Added TO fav List.", Toast.LENGTH_SHORT).show();
+                                SharedPreferences sharedPreference=getSharedPreferences("mypref",MODE_PRIVATE);
+                                SharedPreferences.Editor editor = sharedPreference.edit();
+                                editor.putBoolean("TASK1",true);
+                                editor.commit();
 
                             }
 
@@ -299,7 +338,10 @@ public class View_Place_detail extends AppCompatActivity {
                             finish();
                             startActivity(backIntent);
                             Toast.makeText(View_Place_detail.this, "Place Added TO fav List.", Toast.LENGTH_SHORT).show();
-
+                            SharedPreferences sharedPreference=getSharedPreferences("mypref",MODE_PRIVATE);
+                            SharedPreferences.Editor editor = sharedPreference.edit();
+                            editor.putBoolean("TASK1",true);
+                            editor.commit();
                         }
                     }
